Move worker config to conftest
